+++ src/main/java/vn/flast/repositories/ProviderRepository.java
@@ -20,15 +20,12 @@ package vn.flast.repositories;
 /* có trách nghiệm                                                        */
 /**************************************************************************/
 
-
-
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import vn.flast.models.Provider;
 import java.util.List;
 
-public interface ProviderRepository extends JpaRepository<Provider, Long> {
+public interface ProviderRepository extends JpaRepository<Provider, Integer> {
     @Query("FROM Provider p WHERE p.id IN (:ids)")
-    List<Provider> findByListId(List<Long> ids);
+    List<Provider> findByListId(List<Integer> ids);
 }
